Added a check for cutmark timestamp validity
Checks if the begin of a given cutmark is within the duration of the given tracks. If not, it is removed (i.e. ignored for the rest of the operation).
The reasoning behind this check is that you would rather have incorrectly cutbut publishable tracks than an exception (i.e. if the cut marks are badly automatically generated of the actual length of the tracks is not precisely controllable).