Avoid a potential nullpointer exception in thumbnail generation
One of our clients had a bug in Opencast 14.4 when using the old
editor (the editor integrated in the old admin ui (the admin ui based
on angularjs). When they tried to start processing the changes with
a workflow, this would fail with a nullpointer exception.

This attempts to fix the exception, by setting `tempThumbnail` to
something in case `optPreview.isPresent()` is `false`. Because code
later down the line expects this semi-global variable to be something.

I have no idea if this actually fixes the issue, since our client could
also solve the issue by simply turning off thumbnails in the editor
altogether. Furthermore, I could not reproduce their issue. So the only
thing I could test is that this does not break anything :/